+++ lib/components/confirmar-senha/confirmar-senha.dart
@@ -5,15 +5,17 @@ import 'package:z_components/api/confirmar-senha/i-confirmar-senha-service.dart'
 import 'package:z_components/components/z-alert-dialog.dart';
 import 'package:z_components/config/z-dialog.dart';
 import 'package:z_components/styles/main-style.dart';
-import 'package:z_components/view-model/confirmar-senha-viewmodel.dart';
 
 class ConfirmarSenha {
   String token;
-  String cpf;
   BuildContext context;
 
+<<<<<<< HEAD
   ConfirmarSenha(
       {@required this.token, @required this.cpf, @required this.context});
+=======
+  ConfirmarSenha({@required this.token, @required this.context});
+>>>>>>> 280fe78f041f34f8cb7b06e3054bbf718311c04f
 
   Future<bool> dialogConfirmarSenha() {
     return showDialog(
@@ -70,16 +72,17 @@ class ConfirmarSenha {
     );
   }
 
+<<<<<<< HEAD
   Future<void> _verificarPin(String text) {
     IConfirmarSenhaService iConfirmarSenhaService =
         new ConfirmarSenhaService(token);
+=======
+  Future<void> _verificarPin(String text) async {
+    IConfirmarSenhaService iConfirmarSenhaService = new ConfirmarSenhaService(token);
+>>>>>>> 280fe78f041f34f8cb7b06e3054bbf718311c04f
 
-    ConfirmarSenhaViewModel confimarSenhaViewModel = ConfirmarSenhaViewModel(
-      pin: text,
-      cpf: cpf,
-    );
+    var service = await iConfirmarSenhaService.validarSenha(text);
 
-    var service = iConfirmarSenhaService.validarSenha(confimarSenhaViewModel);
     if (service != null && service == true) {
       Navigator.of(context).pop(true);
     } else {