+++ lib/components/z-user-info/z-user-info-view.dart
@@ -248,6 +248,7 @@ class ZUserInfoView extends IView<ZUserInfo> {
   }
 
   Future<void> escolherImagem(ImageSource source) async {
+    _dialogUtils.showProgressDialog();
     var imagem = await ImagePicker.pickImage(source: source, imageQuality: 70);
 
     if (imagem != null) {
@@ -260,7 +261,7 @@ class ZUserInfoView extends IView<ZUserInfo> {
           imagemPerfil = bytes;
         });
 
-        _arquivoService
+        var idAnexo = await _arquivoService
             .enviarImagem(new ArquivoViewModel(
           nome: "perfil.jpg",
           contentType: "image/jpg",
@@ -268,12 +269,15 @@ class ZUserInfoView extends IView<ZUserInfo> {
           conteudo: base64,
           tamanho: base64.length.toDouble(),
           container: "teste",
-        ))
-            .then((idAnexo) {
-          state.widget.userInfo.idFoto = idAnexo;
-          if (state.widget.onChangeProfileImage != null)
-            state.widget.onChangeProfileImage(base64);
-        });
+        ));
+
+        state.widget.userInfo.idFoto = idAnexo;
+
+        if (state.widget.onChangeProfileImage != null)
+
+          state.widget.onChangeProfileImage(base64);
+
+        _dialogUtils.dismiss();
       }
     }
   }